requirements.txt
Flask==3.0.3
kaggle==1.6.17
supabase-py==0.0.2
python-dotenv==0.21.0